fix: resolve CI test failures and analyzer warnings
- Fix type casting issue in YOLO.predict method for malformed box responses
- Filter out invalid box items instead of converting them to empty maps
- Fix integer division type error in ModelManager download progress calculation

+++ demo_app/lib/services/model_manager.dart
@@ -195,7 +195,7 @@ class ModelManager {
         downloadedBytes += chunk.length;
 
         if (contentLength > 0) {
-          final progress = downloadedBytes / contentLength;
+          final progress = downloadedBytes / contentLength.toDouble();
           onDownloadProgress?.call(progress);
         }
       }
@@ -339,7 +339,7 @@ class ModelManager {
         downloadedBytes += chunk.length;
 
         if (contentLength > 0) {
-          final progress = downloadedBytes / contentLength;
+          final progress = downloadedBytes / contentLength.toDouble();
           onDownloadProgress?.call(progress);
         }
       }
